openenv/tbench2: split the provider-agnostic recipe from the Daytona half
Review feedback on #1710: the module mixed two concerns its own docstring
already kept apart. tb2_task_recipe.py now owns everything Daytona-agnostic
(server layer commands, env-source embedding, task staging, server_cmd,
task.toml reading, /health polling); tb2_task_sandbox.py keeps its name, CLI,
and import surface but is purely the Daytona materialization (declarative
Image build, Resources, create_task_sandbox, bake).

The two symbols that crossed the new module boundary go public with the move:
server_layer_commands and resolve_docker_image.
